完善业务上增加role，删除user的逻辑

+++ src/main/java/com/pollutioncms/service/api/impl/RoleServiceImpl.java
@@ -3,9 +3,14 @@ package com.pollutioncms.service.api.impl;
 import com.github.pagehelper.PageHelper;
 import com.pollutioncms.common.enums.ExceptionEnum;
 import com.pollutioncms.common.exception.DaoException;
+import com.pollutioncms.common.exception.ParamErrorException;
 import com.pollutioncms.module.domain.Role;
+import com.pollutioncms.module.domain.RoleUser;
 import com.pollutioncms.module.mapper.RoleMapper;
+import com.pollutioncms.module.mapper.RoleUserMapper;
+import com.pollutioncms.module.mapper.UserMapper;
 import com.pollutioncms.service.api.RoleService;
+import com.pollutioncms.service.api.UserService;
 import com.pollutioncms.service.dto.RoleDTO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,6 +18,8 @@ import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import tk.mybatis.mapper.entity.Example;
+import tk.mybatis.mapper.util.Sqls;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -30,6 +37,9 @@ public class RoleServiceImpl implements RoleService {
     @Autowired
     private RoleMapper rolemapper;
 
+    @Autowired
+    private RoleUserMapper roleUserMapper;
+
     @Override
     public List<RoleDTO> listRoles(Integer pageNum, Integer count) {
         PageHelper.startPage(pageNum, count);
@@ -37,17 +47,22 @@ public class RoleServiceImpl implements RoleService {
     }
 
     @Override
-    public boolean saveRole(RoleDTO roleDTO) {
-        if (rolemapper.insert(roleDTO.toDO()) != 1) {
-            logger.error("dao operate effect num error,dto:{}",roleDTO);
-            throw new DaoException(ExceptionEnum.DATA_EFFECT_NUM_ERROR);
-        }
+    public boolean saveRole(String parentRoleName,RoleDTO roleDTO) {
+        rolemapper.saveRole(parentRoleName,roleDTO.toDO());
         return true;
 
     }
 
     @Override
     public boolean deleteRole(RoleDTO roleDTO) {
+        int userCount=roleUserMapper.selectCountByExample(Example
+                    .builder(RoleUser.class)
+                    .where(Sqls.custom().andEqualTo("roleId",roleDTO.getId())).build());
+        if(userCount !=0){
+            logger.warn("want delete role has user,dto:{}", roleDTO);
+            throw new ParamErrorException("can't delete role,it has users");
+        }
+
         if (rolemapper.delete(roleDTO.toDO()) != 1){
             logger.error("dao operate effect num error,dto:{}",roleDTO);
             throw new DaoException(ExceptionEnum.DATA_EFFECT_NUM_ERROR);
